+++ R/utils.R
@@ -1,3 +1,3 @@
-BUILD_METADATA_PATTERN <- "^[0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*$"
+BUILD_METADATA_PATTERN <- PRE_RELEASE_IDS_PATTERN <- "^([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*)$"
 
 DEFAULT_ID_LENGTH <- 5L