+++ tests/testthat/test-parse.R
@@ -25,18 +25,6 @@ test_that("parse_pre_release_ids rejects invalid pre-release identifiers", {
   expect_snapshot(parse_pre_release_ids(c("alpha..beta", "..", "--", "1.2.3.")))
 })
 
-test_that("parse_pre_release_ids_impl does not rejects invalid pre-release identifiers", {
-  expect_snapshot(parse_pre_release_ids_impl(
-    c(
-      "alpha..",
-      "..",
-      "--",
-      "1.2.3."
-    ),
-    check = FALSE
-  ))
-})
-
 test_that("Supports more than 5 pre-release identifiers", {
   expect_snapshot(parse_semver("1.2.3-alpha.1.2.3.4.5.6.7.8.9"))
   expect_snapshot(parse_pre_release_ids("alpha.1.2.3.4.5.6.7.8.9"))