+++ tests/testthat/test-smvr.R
@@ -145,7 +145,6 @@ test_that("cast from/to numeric_version", {
 
 test_that("invalid build metadata", {
   expect_snapshot(
-    smvr(1, build = c("foo", "bar.baz", NA, "@foo")),
-    error = TRUE
+    smvr(1, build = c("foo", "bar.baz", NA, "@foo"))
   )
 })