+++ NEWS.md
@@ -5,6 +5,8 @@
 - `new_pre_release_ids()`'s arguments have been changed from `id1, id2, id3, id4, id5` to `...`.
   This allows for more than 5 pre-release identifiers to be specified. (#8)
 - `mark_as_pre_release()`'s `ids` argument no longer has a default value. (#23)
+- The `smvr()` returns `NA` with a warning instead of an error when an invalid value
+  is passed to the `build` argument. (#26)
 
 ## Enhancements